Update env_taxa_correlation.R
Line 59 has a "filtering" on rownames(mt_env) [i.e. the samples] but abund_table was created by otu_table(physeq) giving out an OTUxSamples table and not a SamplexOTU (at least in my case, not doing the compelte pipeline inside microbiomeSeq). Transposing the table can solve the problem and the error arising